backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
sqlalchemy==2.0.21
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.24.1
apscheduler==3.10.4
alembic==1.12.0
python-multipart==0.0.6
cryptography==41.0.4
slowapi==0.1.8
python-json-logger==2.0.7
